requirements.txt
fastapi
uvicorn
python-dotenv
requests
pyjwt
python-multipart
google-generativeai
supabase
huggingface-hub
